% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notionR_parts.R
\name{callAPI}
\alias{callAPI}
\title{Call API}
\usage{
callAPI(id, secret, filters, cursor = NULL)
}
\arguments{
\item{id}{ID of the Notion page or database.}

\item{secret}{Notion API token.}

\item{filters}{A list built with filter operators (see filters) to query database. If NULL will query everything.}
}
\description{
This function is divieded and modified from \code{\link[=getNotionDatabase]{getNotionDatabase()}} and is utilized in other functions.
}
