% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_coordinates.R
\name{dms_to_decimal}
\alias{dms_to_decimal}
\title{Transform DMS coordinates in decimal format}
\usage{
dms_to_decimal(dms, with_letter = "AUTO")
}
\arguments{
\item{dms}{Longitude/Latitude coordinates in DMS format.}

\item{with_letter}{Is the coordinates indicated with letter (N, S, W, E)? Default set to "AUTO", also accept \code{TRUE} or \code{FALSE}.}
}
\value{
A \code{dbl} or a vector of \code{dbl}.
}
\description{
Transform DMS coordinates in decimal format
}
\details{
Vectorized funciton. Input can either be a single character string or a vector of character. DMS coordinates should be separated by \verb{°} or \verb{'} to be able to be recognized.
}
\author{
Zhehao Hu
}
