% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveNotionPage.R
\name{getPage}
\alias{getPage}
\title{Retrieve Metadata of a Notion Page}
\usage{
getPage(id, secret, title_only = FALSE)
}
\arguments{
\item{id}{ID of the Notion page.}

\item{secret}{Notion API token.}

\item{title_only}{If TRUE, only the page title will return.}
}
\value{
Named character vector or character.
}
\description{
GET a Notion page with its metadata.
}
\author{
Zhehao Hu
}
