% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notionR_parts.R
\name{postAPI}
\alias{postAPI}
\title{Call (POST) Notion API}
\usage{
postAPI(secret, id, filters = NULL)
}
\arguments{
\item{secret}{Notion API token.}

\item{id}{ID of the Notion page or database.}

\item{filters}{A list built with filter operators (see filters) to query database. If NULL will query everything.}
}
\value{
as_tibble
}
\description{
Query a database in Notion with desired filters
}
\details{
This function is divieded and modified from \code{\link[=getNotionDatabase]{getNotionDatabase()}}.
}
\author{
Eduardo Flores, Zhehao Hu
}
